v0.2.2
This is a minor release with some additions, but no breaking changes.

New is `toAlpha` and `toAlphaInto` in `lu.conv`; an eager and a lazy template
function respectively that translates an integral into ASCII text, without
having to use `std.format.format` (or `std.format.formattedWrite`).
It can be used in inner loops to slim things down.

Other changes:

* `lu.common.Labeled` now has different names for its members, but
  aliases are in place to make it behave as usual without changes
* CI improvements
* Minor `lu.net.listenFiber` optimisations