v0.5.0
This release removes the lu.net module. It was deemed too specific to
importing projects to warrant being in general purpose lu. Additionally
it pulled the requests dependency, which only made sense if you wanted
to use lu.net (and was baggage otherwise).

As this is obviously a breaking change, this bumps the minor version.

No other significant changes.