v1.1.4
This is a maintenance release with only minor fixes and additions.

The one major change was to split `serialise` up into sub-templates, so
as to reduce the amount of templates being instantiated. One call to
`serialise` now walks through the aggregate object passed and, on a
type-by-type basis, invokes further templates to serialise its members.
This means that for 5 aggregates passed each containing 10 members, the
number of total templates will now be limited to one for each shared
type (plus some `serialise` dispatching overhead), as opposed to
the previous (worst-case) full 5*10.

Compilation memory and template bloat should decrease somewhat, but
general behaviour should otherwise not be affected.