v1.2.4
This release changes how `formatDeltaInto` formats arrays.

Previously it generated one line per array, which led to some
hard-to-visually-parse output like the following assert line of a
`string[8]`;

    assert((someArr == [ "a", "", "c", "", "", "", "", "" ]]),
        someArr.to!string);

This additionally made it brittle in that the assert only held as long
as the arrays kept their original size. A change to the code that made
an `int[3]` array into an `int[5]` would break the expression.

Now one line is generated per (delta) element.

    assert((someArr[0] == "a"), someArr[0]);
    assert((someArr[2] == "c"), someArr[2]);