v3.1.0
This release disables all compiler preview flags in normal `library`
builds. Importing projects may enable them if they so wish.

Added `lu.array.truthTable`. Given an input of numeric values, it
generates a `bool[]` array of a length that accomodates the size of
the input. Indexing into the array with a numeric value yields `true`
if it was in the input set; `false` otherwise. Overloads that generate
static arrays are also included, as well as some that are tailored
specifically for enums.

Care has to be taken when creating truth tables with big numbers,
as the generated array will be of a size accomodating the largest
input value.

`lu.typecons.UnderscoreOpDispatcher` was generalised into `OpDispatcher`.
The former remains but is now implemented using the latter.

* Improved `lu.json.getOrFallback` to be closer to usable
* Added reverse domain functionality to `lu.common.sharedDomains`
* Added some traits to `lu.traits`, mostly to enable the truth table
* Renamed `lu.conv.enumToString` to `lu.conv.toString` for prettier
  UFCS; the original remains as a deprecated compatibility alias
* Removed the copy/pasted `std.traits.getSymbolsByUDA` from `lu.traits`,
  as we now only support versions of Phobos where it is no longer broken
* Deprecated compatibility alias `lu.deltastrings.formatDeltaInto`;
  use `putDelta` instead
* Moves `lu.objmanip.pruneAA` into `lu.array`; a deprecated public
  import remains at the old location
* Made `MutexedAA.aa` private; use `MutexedAA.aaOf` to access it
* Made `lu.string.unenclosed` public