v3.2.0
`lu.common` is now `lu.misc`, but migration should be just simple string
replacements. An old module remains with public imports to the new
location, but it will be removed in a future release.

Added `lu.assert_.assertMultilineEquals` that `assert`s that two strings
of multiple lines are equal, and visualises where lines differ when they
don't. This is useful for tests where you have a large multi-line string
that you want to compare with another, where the difference turns out to
not be immediately obvious.

* Fixed deltastring output of `putDelta` of arrays of strings containing
  characters that should be escaped (and were not)
* Added a simple CI tool for running `dscanner` style checks on the codebase